# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/comfyApp/create": {
            "post": {
                "description": "comfy应用 创建",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ComfyApp"
                ],
                "summary": "comfy应用 创建",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ComfyAppCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回值",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/comfyApp/index": {
            "post": {
                "description": "comfy应用 分页查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ComfyApp"
                ],
                "summary": "comfy应用 分页查询",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ComfyAppIndexRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ComfyAppIndexResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/comfyApp/remove": {
            "post": {
                "description": "comfy应用 删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ComfyApp"
                ],
                "summary": "comfy应用 删除",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ComfyAppRemoveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回值",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/comfyApp/update": {
            "post": {
                "description": "comfy应用 更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ComfyApp"
                ],
                "summary": "comfy应用 更新",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ComfyAppUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回值",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/comfyApp/wsRestart": {
            "post": {
                "description": "comfy应用 ws重启",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ComfyApp"
                ],
                "summary": "comfy应用 ws重启",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ComfyAppRestartWsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回值",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/endpoint/create": {
            "post": {
                "description": "端点 创建",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Endpoint"
                ],
                "summary": "端点 创建",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EndpointCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回值",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/endpoint/index": {
            "post": {
                "description": "端点 分页查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Endpoint"
                ],
                "summary": "端点 分页查询",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EndpointIndexRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.EndpointIndexResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/endpoint/update": {
            "post": {
                "description": "端点 更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Endpoint"
                ],
                "summary": "端点 更新",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EndpointCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回值",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/endpointParam/find": {
            "post": {
                "description": "端点接口参数 查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EndpointParam"
                ],
                "summary": "端点接口参数 查询",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EndpointParamFindRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.EndpointParam"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/endpointParam/set": {
            "post": {
                "description": "端点接口参数 配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EndpointParam"
                ],
                "summary": "端点接口参数 配置",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EndpointParamSetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回值",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Endpoint": {
            "type": "object",
            "properties": {
                "api_json": {
                    "description": "comfyUI api用工作流",
                    "type": "string"
                },
                "app_uid": {
                    "description": "暂时无用，多服务器时可以指定appUid",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "sync_path": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workflow": {
                    "type": "string"
                }
            }
        },
        "model.EndpointParam": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endpoint_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "json_key": {
                    "description": "用于json解析的key",
                    "type": "string"
                },
                "param_key": {
                    "type": "string"
                },
                "param_name": {
                    "type": "string"
                },
                "param_type": {
                    "description": "string, int, float",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "request.ComfyAppCreateRequest": {
            "type": "object",
            "properties": {
                "addr": {
                    "description": "127.0.0.1:8000",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "request.ComfyAppIndexRequest": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "request.ComfyAppRemoveRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "request.ComfyAppRestartWsRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "request.ComfyAppUpdateRequest": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "request.EndpointCreateRequest": {
            "type": "object",
            "properties": {
                "api_json": {
                    "description": "comfy api版本json",
                    "type": "string"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "path": {
                    "description": "路径",
                    "type": "string"
                },
                "workflow": {
                    "description": "comfy 工作流",
                    "type": "string"
                }
            }
        },
        "request.EndpointIndexRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "request.EndpointParamFindRequest": {
            "type": "object",
            "properties": {
                "endpoint_id": {
                    "type": "integer"
                }
            }
        },
        "request.EndpointParamItemRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "json_key": {
                    "type": "string"
                },
                "param_key": {
                    "type": "string"
                },
                "param_name": {
                    "type": "string"
                },
                "param_type": {
                    "description": "string/int/float",
                    "type": "string"
                }
            }
        },
        "request.EndpointParamSetRequest": {
            "type": "object",
            "properties": {
                "endpoint_id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.EndpointParamItemRequest"
                    }
                }
            }
        },
        "response.ComfyAppCreateResponseItem": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pending_num": {
                    "type": "integer"
                },
                "running_num": {
                    "type": "integer"
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "ws_status": {
                    "description": "close,connect",
                    "type": "string"
                }
            }
        },
        "response.ComfyAppIndexResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ComfyAppCreateResponseItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "response.EndpointIndexResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Endpoint"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}